Compare binary causal models in a phylogenetic context.

Compare binary causal models in a phylogenetic context.

phylo_path_binary(models, data, tree, order = NULL, parallel = NULL,
  na.rm = TRUE)

Arguments

models
A list of directed acyclic graphs. These are matrices, typically created with DAG.

data
A data.frame with binary data. Variables can be a numberic vector (with all values either 0 or 1), a factor (with two levels) or character vector (with two unique values).

tree
A phylogenetic tree of class pylo.

order
Causal order of the included variable, given as a character vector. This is used to determine which variable should be the dependent in the dsep regression equations. If left unspecified, the order will be automatically determined. If the combination of all included models is itself a DAG, then the ordering of that full model is used. Otherwise, the most common ordering between each pair of variables is used to create a general ordering.

parallel
An optional vector containing the virtual connection process type for running the chains in parallel (such as "SOCK"). A cluster is create using the parallel package.

na.rm
Should rows that contain missing values be dropped from the data as necessary (with a message)?

Value

A phylopath_binary object, with the following components:

d_sep

for each model a table with seperation statements and statistics.

models

the DAGs

data

the supplied data

tree

the supplied tree

Details

Estimation of relationships between binary variables is built upon ape::binaryPGLMM and stands or falls with the accuracy of that method. Model fitting using this method is particularly slow and the parallel argument can therefore be especially useful.