Data on brain size, life history and vulnerability to extinction

A dataset with continuous variables affecting the conservation Status of mammalian species (the IUCN red list of threatened species).

red_list

Format

An object of class data.frame with 474 rows and 7 columns.

Source

Gonzalez-Voyer A, Gonzalez-Suarez M, Vila C, Revilla E (2016) Larger brain size indirectly increases vulnerability to extinction in mammals. Evolution 70:1364-1375. doi: 10.1111/evo.12943

Details

It includes the following variables: brain size (Br), body size (B), gestation period (G), litter size (L), weening age (W), population density (P) and vulnerability to extinction (Status).