Extract and average the best supported models from a phylogenetic path analysis.

Extract and average the best supported models from a phylogenetic path analysis.

average(phylopath, cut_off = 2, avg_method = "conditional", ...)

Arguments

phylopath
An object of class phylopath.

cut_off
The CICc cut-off used to select the best models. Use Inf to average over all models. Use the best() function to only use the top model, or choice() to select any single model.

avg_method
Either "full" or "conditional". The methods differ in how they deal with averaging a path coefficient where the path is absent in some of the models. The full method sets the coefficient (and the variance) for the missing paths to zero, meaning paths that are missing in some models will shrink towards zero. The conditional method only averages over models where the path appears, making it more sensitive to small effects. Following von Hardenberg & Gonzalez-Voyer 2013, conditional averaging is set as the default. Also see MuMIn::model.avg().

...
Arguments to pass to phylolm::phylolm and phylolm::phyloglm. If you specified options in the original phylo_path call you don't need to specify them again.

Value

An object of class fitted_DAG.

Examples

  candidates <- define_model_set(
    A = NL ~ BM,
    B = NL ~ LS,
    .common = c(LS ~ BM, DD ~ NL)
  )
  p <- phylo_path(candidates, rhino, rhino_tree)
  summary(p)

#>   model k q     C       p CICc delta_CICc        l        w
#> A     A 3 7  2.65 0.85113 17.9        0.0 1.00e+00 1.00e+00
#> B     B 3 7 22.11 0.00115 37.3       19.5 5.94e-05 5.94e-05

  # Models A and B have similar support, so we may decide to take
  # their average.

  avg_model <- average(p)
  # Print the average model to see coefficients, se and ci:
  avg_model

#> $coef
#>    BM        LS        NL        DD
#> BM  0 0.4973937 0.4303443 0.0000000
#> LS  0 0.0000000 0.0000000 0.0000000
#> NL  0 0.0000000 0.0000000 0.6285344
#> DD  0 0.0000000 0.0000000 0.0000000
#> attr(,"class")
#> [1] "matrix" "DAG"   
#> 
#> $se
#>    BM         LS         NL         DD
#> BM  0 0.08934185 0.08868688 0.00000000
#> LS  0 0.00000000 0.00000000 0.00000000
#> NL  0 0.00000000 0.00000000 0.08006703
#> DD  0 0.00000000 0.00000000 0.00000000
#> 
#> $lower
#>    BM        LS        NL        DD
#> BM  0 0.3222869 0.2565212 0.0000000
#> LS  0 0.0000000 0.0000000 0.0000000
#> NL  0 0.0000000 0.0000000 0.4716059
#> DD  0 0.0000000 0.0000000 0.0000000
#> 
#> $upper
#>    BM        LS        NL        DD
#> BM  0 0.6725005 0.6041674 0.0000000
#> LS  0 0.0000000 0.0000000 0.0000000
#> NL  0 0.0000000 0.0000000 0.7854629
#> DD  0 0.0000000 0.0000000 0.0000000
#> 
#> attr(,"class")
#> [1] "fitted_DAG"

  
# NOT RUN {
  # Plot to show the weighted graph:
  plot(avg_model)

  # One can see that an averaged model is not necessarily a DAG itself.
  # This model actually has a path in two directions.

  # Note that coefficients that only occur in one of the models become much
  # smaller when we use full averaging:

  coef_plot(avg_model)
  coef_plot(average(p, method = 'full'))

# }