Compare causal models in a phylogenetic context.

Compare causal models in a phylogenetic context.

phylo_path(models, data, tree, cor_fun = ape::corPagel, order = NULL,
  parallel = NULL, na.rm = TRUE)

Arguments

models
A list of directed acyclic graphs. These are matrices, typically created with DAG.

data
A data.frame with data.

tree
A phylogenetic tree of class pylo.

cor_fun
A function that creates a corStruct object, typically one of the cor functions from the ape package, such as corBrownian, corPagel etc.

order
Causal order of the included variable, given as a character vector. This is used to determine which variable should be the dependent in the dsep regression equations. If left unspecified, the order will be automatically determined. If the combination of all included models is itself a DAG, then the ordering of that full model is used. Otherwise, the most common ordering between each pair of variables is used to create a general ordering.

parallel
An optional vector containing the virtual connection process type for running the chains in parallel (such as "SOCK"). A cluster is create using the parallel package.

na.rm
Should rows that contain missing values be dropped from the data as necessary (with a message)?

Value

A phylopath object, with the following components:

d_sep

for each model a table with seperation statements and statistics.

models

the DAGs

data

the supplied data

tree

the supplied tree

cor_fun

the employed correlation structure

Examples

  #see vignette('intro_to_phylopath') for more details
  candidates <- list(A = DAG(LS ~ BM, NL ~ BM, DD ~ NL),
                     B = DAG(LS ~ BM, NL ~ LS, DD ~ NL))
  p <- phylo_path(candidates, rhino, rhino_tree)

#>    |++++++++++                                        | 20% ~00s             |++++++++++++++++++++                              | 40% ~00s             |++++++++++++++++++++++++++++++                    | 60% ~00s             |++++++++++++++++++++++++++++++++++++++++         | 80% ~00s             |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 00s


  # Printing p gives some general information:
  p

#> 
#> A phylogenetic path analysis
#> 
#>   Evaluated for these models: A B 
#> 
#>   Containing 6 phylogenetic regressions.

  # And the summary gives statistics to compare the models:
  summary(p)

#>   model k q      C     p   CICc delta_CICc l w
#> 1     A 3 7  4.736 0.578 19.953      0.000 1 1
#> 2     B 3 7 24.116 0.000 39.334     19.381 0 0