Compare causal models in a phylogenetic context.

Continuous variables are modeled using phylolm::phylolm, while binary traits are modeled using phylolm::phyloglm.

phylo_path(model_set, data, tree, model = "lambda",
  method = "logistic_MPLE", order = NULL, parallel = NULL, na.rm = TRUE,
  ...)

Arguments

model_set
A list of directed acyclic graphs. These are matrices, typically created with define_model_set.

data
A data.frame with data. If you have binary variables, make sure they are either character values or factors!

tree
A phylogenetic tree of class phylo.

model
The evolutionary model used for the regressions on continuous variables. See phylolm::phylolm for options and details. Defaults to Pagel's lambda model

method
The estimation method for the binary models. See phylolm::phylolm for options and details. Defaults to logistic MPLE.

order
Causal order of the included variable, given as a character vector. This is used to determine which variable should be the dependent in the dsep regression equations. If left unspecified, the order will be automatically determined. If the combination of all included models is itself a DAG, then the ordering of that full model is used. Otherwise, the most common ordering between each pair of variables is used to create a general ordering.

parallel
An optional vector containing the virtual connection process type for running the chains in parallel (such as "SOCK"). A cluster is create using the parallel package.

na.rm
Should rows that contain missing values be dropped from the data as necessary (with a message)?

...
Arguments passed on to phylolm:

lower.bound: optional lower bound for the optimization of the phylogenetic model parameter.

upper.bound: optional upper bound for the optimization of the phylogenetic model parameter.

starting.value: optional starting value for the optimization of the phylogenetic model parameter.

measurement_error: a logical value indicating whether there is measurement error sigma2_error (see Details).

Arguments passed on to phyloglm:

btol: bound on the linear predictor to bound the searching space.

log.alpha.bound: bound for the log of the parameter alpha.

start.beta: starting values for beta coefficients.

start.alpha: starting values for alpha (phylogenetic correlation).

Value

A phylopath object, with the following components:

d_sep

for each model a table with separation statements and statistics.

model_set

the DAGs

data

the supplied data

tree

the supplied tree

model

the employed model of evolution in phylolm

method

the employed method in phyloglm

dots

any additional arguments given, these are passed on to downstream functions

warnings

any warnings generate by the models

Examples

  #see vignette('intro_to_phylopath') for more details
  candidates <- list(A = DAG(LS ~ BM, NL ~ BM, DD ~ NL),
                     B = DAG(LS ~ BM, NL ~ LS, DD ~ NL))
  p <- phylo_path(candidates, rhino, rhino_tree)

  # Printing p gives some general information:
  p

#> A phylogenetic path analysis, on the variables:
#> 	Continuous:	 BM NL DD LS 
#> 	Binary:		  
#> 
#>  Evaluated for these models: A B 
#> 
#>  Containing 6 phylogenetic regressions, of which 5 unique

  # And the summary gives statistics to compare the models:
  summary(p)

#>   model k q      C     p   CICc delta_CICc l w
#> 1     A 3 7  2.652 0.851 17.869      0.000 1 1
#> 2     B 3 7 22.114 0.001 37.332     19.463 0 0